Introduce common test module
